Split ModuliTests
